+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/projectoverview/CountSamplesPresenter.groovy
@@ -36,6 +36,16 @@ class CountSamplesPresenter implements CountSamplesOutput {
         }
     }
 
+    @Override
+    void countedPassingQcSamples(String projectCode, int allSamples, int passedQcSamples) throws OutputException {
+        try {
+            StatusCount statusCount = new StatusCount(projectCode, Status.SAMPLE_QC_PASS, passedQcSamples, allSamples)
+            statusCountResourceService.addToResource(statusCount)
+        } catch (Exception e) {
+            throw new OutputException(e.getMessage())
+        }
+    }
+
     /**
      * To be called after successfully counting samples for the provided code.
      * @param number of all samples and samples that have been received by the lab