+++ sample-tracking-status-overview-domain/src/main/groovy/life/qbic/business/samples/count/CountSamples.groovy
@@ -42,6 +42,7 @@ class CountSamples implements CountSamplesInput{
       this.projectCode = projectCode
 
       countReceivedSamples()
+      countPassingQcSamples()
       countFailedQcSamples()
       countFinishedLibraryPrepSamples()
       countAvailableDataSamples()
@@ -56,6 +57,11 @@ class CountSamples implements CountSamplesInput{
   private void countReceivedSamples(){
     int receivedAmount = countSamplesFromStatus(Status.SAMPLE_RECEIVED)
     output.countedReceivedSamples(projectCode,totalSampleCount,receivedAmount)
+  }
+
+  private void countPassingQcSamples(){
+    int passedQc = countSamplesFromStatus(Status.SAMPLE_QC_PASS)
+    output.countedPassingQcSamples(projectCode,totalSampleCount,passedQc)
   }
 
   private void countFailedQcSamples(){